# api/SWAGGER-API.yaml
openapi: 3.0.0
info:
  title: Docs Api Monopatines
  description: Utilizacion de api de uso y gestion de app de monopatines
  version: "2.0"
servers:
- url: https://virtserver.swaggerhub.com/LEOPOLDOVILLAO14/tpe-arquitecturas/2.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8080
  description: Swagger
...
paths:
...
  /monopatines/:
    get:
      tags:
      - Monopatines
      summary: Obtiene la lista de todos los monopatines creados.
      description: Ideal para listar todos los monopatines.
      responses:
        "200":
          description: Lista de monopatines obtenida con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Monopatin'
    post:
      tags:
      - Monopatines
      summary: Crea un monopatin disponible para viajar.
      description: "Ideal para crear monopatines, asignarlos a paradas para que puedan hacer viajes."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonoDTO'
        required: true
      responses:
        "201":
          description: Monopatin creado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonoDTO'
  /monopatines/{id}:
    get:
      tags:
      - Monopatines
      summary: "Obtiene un monopatin asociado a su {id}."
      description: Ideal para ver detalles de un monopatin en particular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Id del monopatin a buscar.
      responses:
        "200":
          description: Monopatin obtenido con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monopatin'
    put:
      tags:
      - Monopatines
      summary: "Modifica un monopatin asociado a su {id}."
      description: Ideal para modificar propiedades de un monopatin en particular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Id del monopatin a modificar.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMonoDTO'
        required: true
      responses:
        "200":
          description: Monopatin modificado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateMonoDTO'
    delete:
      tags:
      - Monopatines
      summary: "Elimina un monopatin asociado a su {id}."
      description: Ideal para eliminar un monopatin en particular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Id del monopatin a eliminar.
      responses:
        "200":
          description: Monopatin eliminado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monopatin'
  /monopatines/mas-viajes:
    get:
      tags:
      - Monopatines
      summary: Informe de monopatines por año y cantidad de viajes.
      description: Obtiene una lista que muestra el id de cada monopatin y el total de viajes en el año pasado por parametro. Se puede establecer mostrar los monopatines cuya cantidad de viajes sea mayor a el parametro 'minViaje'.
      parameters:
      - name: anio
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Año en el cual se desea saber la cantidad de viajes de los monopatines.
      - name: minViaje
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Minimo de viajes que deben tener los monopatines para incluirse en el listado.
      responses:
        "200":
          description: Informe de viajes por año y minimo de viajes obtenido con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTravelInYearDTO'
...
  /monopatines/utilizacion/{id}:
    get:
      tags:
      - Monopatines
      summary: "Informe que muestra total de viajes, tiempo viajado, tiempo en pausa del monopatin asociado al {id}."
      description: "Util para mostrar cuantos viajes hizo un monopatin, el tiempo total que estuvo viajando y opcionalmente el tiempo total que estuvo en pausa."
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Id del monopatin solicitado.
      - name: conPausas
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: boolean
          description: Indica true si requiere el tiempo total de pausas.
      responses:
        "200":
          description: Informe generado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
  /monopatines/disponibilidad/{id}:
    put:
      tags:
      - Monopatines
      summary: "Alterna la disponibilidad del monopatin asociado a su {id}."
      description: "Util para cambiar la disponibilidad de un monopatin en particular. Si el monopatin esta DISPONIBLE , lo cambia a NO DISPONIBLE. Si el monopatin esta en NO DISPONIBLE, lo cambia a DISPONIBLE."
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Id del monopatin a actualizar la disponibilidad.
      responses:
        "200":
          description: Disponibilidad del monopatin actualizada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateMonoDTO'
  /monopatines/ubicacion/{id}:
    put:
      tags:
      - Monopatines
      summary: "Actualiza la ubicacion del monopatin que está en viaje asociado a su {id}."
      description: Util para actualizar la ubicacion cuando un monopatin se mueve en un viaje de un lado a otro o de una parada a otra.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Id del monopatin a actualizar ubicacion.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ubicacion_id_body'
        required: true
      responses:
        "200":
          description: Ubicacion del monopatin actualizada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KmsLocationDTO'
...
  /viajes/:
    get:
      tags:
      - Viajes
      summary: Obtiene una lista de todos los viajes creados.
      description: Util para listar todos los viajes realizados y en curso.
      responses:
        "200":
          description: Lista de viajes obtenida con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelDTO'
    post:
      tags:
      - Viajes
      summary: Crea un nuevo viaje en monopatin.
      description: Util para dar inicio a un viaje en monopatin por parte de un usuario.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitTravelDTO'
        required: true
      responses:
        "201":
          description: Viaje iniciado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTravelDTO'
  /viajes/pausar/{id}:
    put:
      tags:
      - Viajes
      summary: "Pone en pausa el viaje que coincide con su {id}."
      description: Util para pausar un viaje actualmente activo.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id del viaje a poner en pausa.
      responses:
        "200":
          description: Viaje puesto en pausa con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PauseResponseDTO'
  /viajes/reanudar/{id}:
    put:
      tags:
      - Viajes
      summary: "Reanuda un viaje que coincide con el {id} y fue puesto en pausa anteriormente."
      description: "Util para reanudar un viaje que fue puesto en pausa. Se cobrará por el tiempo desde iniciado el viaje ó desde el final de la ultima pausa hasta el momento en que se reanuda el viaje. Si pasaron mas de 15 minutos en pausa, se cambiará la tarifa activa a una adicional para el resto del viaje."
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id del viaje a reanudar.
      responses:
        "200":
          description: Viaje reanudado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinueTravelDTO'
  /viajes/finalizar/{id}:
    put:
      tags:
      - Viajes
      summary: "Finaliza un viaje en curso que coincide con su {id}."
      description: "Util para finalizar un viaje que estaba siendo desarrollado. El monopatin deberá estar en la ubicacion de la parada que se marco como destino. Debita dinero acorde al tiempo que no haya sido cobrado, por ejemplo minutos de diferencia entre el final de una pausa y la hora de finalizacion Ó cobrará el total del tiempo de viaje si no se hizo pausa en ningun momento."
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id del viaje que se desea finalizar.
      responses:
        "200":
          description: Viaje finalizado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloseTravelDTO'
  /viajes/informe/monopatines:
    get:
      tags:
      - Viajes
      summary: Mismo efecto `/monopatines/mas-viajes`.
      description: Este endpoint aporta el mismo informe ubicado en `/monopatines/mas-viajes`.
      parameters:
      - name: anio
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Año en el cual se desea saber la cantidad de viajes de los monopatines.
      - name: minViaje
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Minimo de viajes que deben tener los monopatines para incluirse en el listado.
      responses:
        "200":
          description: Informe de viajes por año y minimo de viajes obtenido con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTravelInYearDTO'
  /viajes/informe/facturado:
    get:
      tags:
      - Viajes
      summary: Informe del total facturado entre un rango de meses de un año.
      description: "Genera un informe del total facturado de viajes en un rango de meses de un año. Recibe por parametros el numero de mes inicial, numero de mes final y opcionalmente el año. Si no se incluye el año, el analisis será sobre el corriente año."
      parameters:
      - name: mesInicial
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Numero de mes inicial.
      - name: mesFinal
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Numero de mes final.
      - name: anio
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: number
          description: Numero de año de analisis.
      responses:
        "200":
          description: Informe generado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTotalBilledDTO'
...
components:
  schemas:
...
    Point:
      type: object
      properties:
        x:
          type: number
          example: 5
        "y":
          type: number
          example: 5
    MonoDTO:
      required:
      - id_stop
      - isAvailable
      - kms
      type: object
      properties:
        isAvailable:
          type: boolean
          example: true
        id_stop:
          type: number
          description: Id de la parada la cual estara el monopatin.
          example: 2
        kms:
          type: number
          description: kilometros actuales del monopatin.
          example: 280
    StopDTO:
      type: object
      properties:
        id_stop:
          type: number
          example: 2
        name:
          type: string
          example: Parada zona centro
        adress:
          type: string
          example: Avenida libertador 2018
        location:
          $ref: '#/components/schemas/Point'
    Monopatin:
      type: object
      properties:
        id_monopatin:
          type: string
          example: asSJDHSskjdha-23872DbCCd
        isAvailable:
          type: boolean
          example: true
        location:
          $ref: '#/components/schemas/Point'
        kms:
          type: number
          example: 280
        stop_assign:
          $ref: '#/components/schemas/StopDTO'
    UpdateMonoDTO:
      required:
      - isAvailable
      - stopAssign
      type: object
      properties:
        isAvailable:
          type: boolean
          example: true
        stopAssign:
          $ref: '#/components/schemas/StopDTO'
    ReportTravelInYearDTO:
      type: object
      properties:
        id_monopatin:
          type: string
          example: ASJDHqwueySGH7262dUY=eD
        totalTravelsInYear:
          type: number
          example: 50
...
    KmsLocationDTO:
      type: object
      properties:
        newLocation:
          $ref: '#/components/schemas/Point'
        newKms:
          type: number
          example: 287
...
    InitTravelDTO:
      required:
      - id_account
      - id_monopatin
      - id_stop_end
      - id_user
      type: object
      properties:
        id_user:
          type: number
          description: Id de usuario que da comienzo al viaje.
          example: 15
        id_account:
          type: number
          description: Id de cuenta asociada al usuario que da comienzo al viaje.
          example: 2
        id_monopatin:
          type: string
          description: Id del monopatin que va hacer uso el usuario.
          example: QUWETsgE2YUEUW873=edRT
        id_stop_end:
          type: number
          description: Id de la parada en la cual va a finalizar el viaje.
          example: 2
    ResponseDebitDTO:
      type: object
      properties:
        description:
          type: string
          example: "Debito automatico de $[MONTO] por [MOTIVO]."
        total_debit:
          type: number
          example: 500
    CreateTravelDTO:
      type: object
      properties:
        id_travel:
          type: number
          example: 3
        initDate:
          type: string
          example: 2024-11-02
        paymentDetail:
          $ref: '#/components/schemas/ResponseDebitDTO'
    Tarifa:
      type: object
      properties:
        id_fee:
          type: number
          example: 3
        normalFee:
          type: number
          example: 500
        additionalFee:
          type: number
          example: 1500
        dateValidity:
          type: string
          example: 2024-10-01
    TravelDTO:
      type: object
      properties:
        id_travel:
          type: number
          example: 3
        id_user:
          type: number
          example: 18
        id_account:
          type: number
          example: 2
        id_monopatin:
          type: string
          example: QUWETsgE2YUEUW873=edRT
        id_stop_init:
          type: number
          example: 1
        id_stop_end:
          type: number
          example: 2
        date_init:
          type: string
          example: 2024-10-03 15:20
        date_end:
          type: string
        kmsMonoInit:
          type: number
          example: 287
        kmsTraveled:
          type: number
          example: 15
        time:
          type: number
          description: Tiempo total del viaje en minutos.
          example: 25
        state:
          type: boolean
          example: true
        total_pay:
          type: number
          example: 500
        isPause:
          type: boolean
          example: false
        fee:
          $ref: '#/components/schemas/Tarifa'
        feeActive:
          type: number
          example: 500
        quantityEndPauses:
          type: number
          example: 0
    PauseResponseDTO:
      type: object
      properties:
        id_travel:
          type: number
          example: 3
        initPause:
          type: string
          example: 2024-11-03 15:35
    ContinueTravelDTO:
      type: object
      properties:
        minutesCharged:
          type: number
          description: Minutos cobrados (puede ser desde el inicio del viaje o desde el fin de la ultima pausa).
          example: 20
        totalPauseInMin:
          type: number
          description: Duración de la pausa que finalizó expresada en minutos.
          example: 8
        debitDetail:
          $ref: '#/components/schemas/ResponseDebitDTO'
    CloseTravelDTO:
      type: object
      properties:
        id_travel:
          type: number
          example: 3
        totalPay:
          type: number
          example: 2500
        kmsTraveled:
          type: number
          example: 12
        endDate:
          type: string
          example: 2024-11-03 15:55
        debitEnd:
          $ref: '#/components/schemas/ResponseDebitDTO'
    ReportTotalBilledDTO:
      type: object
      properties:
        monthInit:
          type: number
          example: 2
        monthEnd:
          type: number
          example: 11
        year:
          type: number
          example: 2024
        totalInPesos:
          type: number
          example: 25700000
...
    ReportUsageDTO:
      type: object
      properties:
        id_monopatin:
          type: string
          example: QUWETsgE2YUEUW873=edRT
        total_kms:
          type: number
          example: 467
        total_min_traveled:
          type: number
          example: 37465
        total_travels:
          type: number
          example: 150
    ReportUsagePauseDTO:
      allOf:
      - $ref: '#/components/schemas/ReportUsageDTO'
      - type: object
        properties:
          total_min_pause:
            type: number
            example: 180
    inline_response_200:
      oneOf:
      - $ref: '#/components/schemas/ReportUsagePauseDTO'
      - $ref: '#/components/schemas/ReportUsageDTO'
    ubicacion_id_body:
      required:
      - location
      properties:
        location:
          $ref: '#/components/schemas/Point'
...